+++ lib/common_test.go
@@ -4,6 +4,7 @@ import (
 	"fmt"
 	"log"
 	"os"
+	"path/filepath"
 )
 
 func checkFileExist(file string) bool {
@@ -40,4 +41,41 @@ func createDirIfNotExist(dir string) {
 			panic(err)
 		}
 	}
+}
+
+func cleanUp(path string) {
+	removeContents(path)
+	removeFiles(path)
+}
+
+func removeFiles(src string) {
+	files, err := filepath.Glob(src)
+	if err != nil {
+
+		panic(err)
+	}
+	for _, f := range files {
+		if err := os.Remove(f); err != nil {
+			panic(err)
+		}
+	}
+}
+
+func removeContents(dir string) error {
+	d, err := os.Open(dir)
+	if err != nil {
+		return err
+	}
+	defer d.Close()
+	names, err := d.Readdirnames(-1)
+	if err != nil {
+		return err
+	}
+	for _, name := range names {
+		err = os.RemoveAll(filepath.Join(dir, name))
+		if err != nil {
+			return err
+		}
+	}
+	return nil
 }
