+++ lib/download_test.go
@@ -6,7 +6,6 @@ import (
 	"net/url"
 	"os"
 	"os/user"
-	"path/filepath"
 	"testing"
 
 	lib "github.com/warren-veerasingam/terraform-switcher/lib"
@@ -159,41 +158,4 @@ func TestURLValid(t *testing.T) {
 	} else {
 		t.Logf("Valid URL from %v", url)
 	}
-}
-
-func cleanUp(path string) {
-	removeContents(path)
-	removeFiles(path)
-}
-
-func removeFiles(src string) {
-	files, err := filepath.Glob(src)
-	if err != nil {
-
-		panic(err)
-	}
-	for _, f := range files {
-		if err := os.Remove(f); err != nil {
-			panic(err)
-		}
-	}
-}
-
-func removeContents(dir string) error {
-	d, err := os.Open(dir)
-	if err != nil {
-		return err
-	}
-	defer d.Close()
-	names, err := d.Readdirnames(-1)
-	if err != nil {
-		return err
-	}
-	for _, name := range names {
-		err = os.RemoveAll(filepath.Join(dir, name))
-		if err != nil {
-			return err
-		}
-	}
-	return nil
 }
